requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.7.1
Django==5.0.1
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-phonenumber-field==7.3.0
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
josepy==1.14.0
l18n==2021.3
mozilla-django-oidc==4.0.0
openpyxl==3.1.2
phonenumbers==8.13.30
pillow==10.2.0
pillow-heif==0.14.0
psycopg2==2.9.9
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
transliterate==1.10.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uWSGI==2.0.24
wagtail==5.2.3
webencodings==0.5.1
Willow==1.6.3
